requirements.txt
celery==3.1.17
cssselect==0.9.1
Django==1.7.1
django-secure==1.0.1
dj-database-url==0.3.0
django-cors-headers==0.13
django-filter==0.8
django-jsonfield==0.9.13
django-model-utils==2.2.0
djangorestframework==3.0.1
iron-celery==0.4.3
lxml==3.4.1
psycopg2==2.5.4
purl==1.0
raven==5.1.1
requests==2.4.3
